pyproject.toml
[project]
name = "robo-eproc"
version = "0.3.0"
description = "Automação de processos do Eproc TJTO"
authors = [{name = "Maicon Dener"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "playwright>=1.44.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "pydantic-settings>=2.2.1",
    "pyotp>=2.9.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",
    "requests>=2.32.0",
    "google-api-python-client>=2.155.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "black>=24.0.0"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "Q"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
